+++ src/xc_integrator/local_work_driver/host/obara_saika/src/integral_1.h
@@ -3,7 +3,7 @@
 
 #include "integral_1.h"
 
-void integral_1(int npts,
+void integral_1(size_t npts,
                shells shellA,
                point *points,
                double *Xi,
