+++ src/xc_integrator/local_work_driver/host/obara_saika/src/integral_4_3.h
@@ -3,7 +3,7 @@
 
 #include "integral_4_3.h"
 
-void integral_4_3(int npts,
+void integral_4_3(size_t npts,
                   shells shellA,
                   shells shellB,
                   point *points,
